+++ src/components/EventsFinder.jsx
@@ -1,20 +1,57 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React, { Component } from 'react';
+import axios from 'axios';
+// import PropTypes from 'prop-types';
 
-function EventsFinder({ eventsImport }) {
-  return (
-    <div>
-      <p>{eventsImport.title}</p>
-      <p>{eventsImport.geometry[0].coordinates}</p>
-    </div>
-  );
+const options = {
+  method: 'GET',
+  url: 'https://eonet.sci.gsfc.nasa.gov/api/v3/events',
+  params: { status: 'open' },
+  headers: { 'Content-Type': 'application/json' },
+};
+
+class EventsFinder extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      eventsList: [],
+    };
+    this.getEvents = this.getEvents.bind(this);
+  }
+
+  componentDidMount() {
+    this.getEvents();
+  }
+
+  getEvents() {
+    axios
+      .request(options)
+      .then((response) => response.data.events)
+      .then((events) => {
+        this.setState({
+          eventsList: events,
+        });
+      });
+  }
+
+  render() {
+    const { eventsList } = this.state;
+    return (
+      <div>
+        <button type="button" onClick={this.getEvents}>
+          Get events
+        </button>
+        <p>{eventsList.title}</p>
+        <p>{eventsList.geometry[0].coordinates}</p>
+      </div>
+    );
+  }
 }
 
-EventsFinder.propTypes = {
-  eventsImport: PropTypes.shape.isRequired,
-  // title: PropTypes.string.isRequired,
-  // geometry: PropTypes.string.isRequired,
-  // coordinates: PropTypes.string.isRequired,
-};
+// EventsFinder.propTypes = {
+//   eventsList: PropTypes.shape.isRequired,
+//   title: PropTypes.string.isRequired,
+//   geometry: PropTypes.string.isRequired,
+//   coordinates: PropTypes.string.isRequired,
+// };
 
 export default EventsFinder;
